requirements.txt
PySimpleGUI
GitPython
requests
appdirs
